Initial CLMatrix class implementation.
