+++ src/aspire/abinitio/__init__.py
@@ -5,6 +5,7 @@ from .commonline_utils import (
     g_sync,
 )
 from .commonline_base import CLOrient3D
+from .commonline_matrix import CLMatrix
 from .commonline_sdp import CommonlineSDP
 from .commonline_lud import CommonlineLUD
 from .commonline_irls import CommonlineIRLS
