+++ src/aspire/abinitio/commonline_sdp.py
@@ -4,14 +4,14 @@ import cvxpy as cp
 import numpy as np
 from scipy.sparse import csr_array
 
-from aspire.abinitio import CLOrient3D
+from aspire.abinitio import CLMatrix
 from aspire.utils import nearest_rotations
 from aspire.utils.matlab_compat import stable_eigsh
 
 logger = logging.getLogger(__name__)
 
 
-class CommonlineSDP(CLOrient3D):
+class CommonlineSDP(CLMatrix):
     """
     Class to estimate 3D orientations using semi-definite programming.
